Factor out field
